requirements.txt
flax==0.4.1
jax==0.4.6
isort==5.9.3
jinja2>=2.11
numpy==1.20.1  # sweet spot between tensorflow and torch packages
pylint==2.11.0
scipy
toml==0.10.2
tqdm==4.64.0
wget==3.2
optax==0.1.2
tifffile
pandas
gpustat 
torch 
torchmetrics==0.6.2
scipy 
pandas 
torchvision 
torchaudio 
numba
Pillow
pytorch-lightning
librosa
datasets
tensorflow
tensorflow-datasets
mat73
transformers
wandb
einops
yapf 
typing-extensions 
fsspec 
flatten-dict